# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Proof Platform API",
    "version": "1.0.0",
    "description": "API documentation for Proof Platform - Where Execution Speaks",
    "contact": {
      "name": "API Support",
      "email": "support@proof.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://api.proof.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sb-access-token"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        },
        "required": ["error"]
      },
...
      "Hypothesis": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "startup_id": { "type": "string" },
          "student_id": { "type": "string" },
          "si_clause": { "type": "string" },
          "entonces_clause": { "type": "string" },
          "porque_clause": { "type": "string" },
          "validation_method": { "type": "string" },
          "target_metric": { "type": "string" },
          "success_criteria": { "type": "string" },
          "status": { "type": "string", "enum": ["draft", "validated", "invalidated"] },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "TalentScore": {
        "type": "object",
        "properties": {
          "totalScore": { "type": "number" },
          "executionScore": { "type": "number" },
          "engagementScore": { "type": "number" },
          "tractionScore": { "type": "number" },
          "collaborationScore": { "type": "number" },
          "innovationScore": { "type": "number" },
          "category": { "type": "string", "enum": ["high_performer", "on_track", "at_risk"] },
          "trend": { "type": "string", "enum": ["up", "down", "stable"] }
        }
      }
    }
  },
  "paths": {
...
    "/api/hypothesis/generate": {
      "post": {
        "summary": "Generate hypotheses using AI",
        "tags": ["Hypothesis"],
        "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "startupId": { "type": "string" },
                  "problema": { "type": "string" },
                  "usuario": { "type": "string" },
                  "por_que_ahora": { "type": "string" },
                  "riesgo": { "type": "string" }
                },
                "required": ["startupId", "problema", "usuario", "por_que_ahora", "riesgo"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Hypotheses generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "hypotheses": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Hypothesis" }
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/api/talent/score": {
      "get": {
        "summary": "Get talent score for student/startup",
        "tags": ["Talent"],
        "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }],
        "parameters": [
          {
            "name": "studentId",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "startupId",
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Talent score retrieved",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TalentScore" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    }
  }
}